DoubleQuote checker
Funcionalidade de checar se temos aspas nao fechadas na string de argumento passada.
